requirements.txt

fastapi==0.104.0
uvicorn==0.23.2
sqlalchemy==2.0.36
psycopg[binary]==3.2.6
pydantic-settings==2.0.3
google-generativeai==0.3.1
httpx==0.25.0
python-dotenv==1.0.0
